+++ console/src/components/Services/Data/TableBrowseRows/ViewRows.js
@@ -45,6 +45,7 @@ import {
   findTable,
   getRelationshipRefTable,
   dataSource,
+  currentDriver,
 } from '../../../../dataSources';
 import { updateSchemaInfo } from '../DataActions';
 import {
@@ -928,6 +929,57 @@ const ViewRows = props => {
       }
     };
 
+    const PaginationWithOnlyNav = () => {
+      const newPage = curFilter.offset / curFilter.limit;
+      return (
+        <div className={`row`} style={{ maxWidth: '500px' }}>
+          <div className="col-xs-2">
+            <button
+              className="btn"
+              onClick={() => handlePageChange(newPage - 1)}
+              disabled={curFilter.offset === 0}
+            >
+              prev
+            </button>
+          </div>
+          <div className="col-xs-4">
+            <select
+              value={curFilter.limit}
+              onChange={e => {
+                e.persist();
+                handlePageSizeChange(parseInt(e.target.value, 10) || 10);
+              }}
+              className="form-control"
+            >
+              <option disabled value="">
+                --
+              </option>
+              <option value={5}>5 rows</option>
+              <option value={10}>10 rows</option>
+              <option value={20}>20 rows</option>
+              <option value={25}>25 rows</option>
+              <option value={50}>50 rows</option>
+              <option value={100}>100 rows</option>
+            </select>
+          </div>
+          <div className="col-xs-2">
+            <button
+              className="btn"
+              onClick={() => handlePageChange(newPage + 1)}
+              disabled={curRows.length === 0}
+            >
+              next
+            </button>
+          </div>
+        </div>
+      );
+    };
+
+    const mssqlProps = {};
+    if (currentDriver === 'mssql') {
+      mssqlProps.PaginationComponent = PaginationWithOnlyNav;
+    }
+
     return (
       <DragFoldTable
         className="dataTable -highlight -fit-content"
@@ -959,6 +1011,7 @@ const ViewRows = props => {
         }
         defaultReorders={columnsOrder}
         showPagination={!shouldHidePagination}
+        {...mssqlProps}
       />
     );
   };