+++ console/src/dataSources/services/mssql/index.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { DataSourcesAPI } from '../..';
 import { TableColumn, Table } from '../../types';
+import { getTableRowRequest } from './utils';
 
 const permissionColumnDataTypes = {
   character: [
@@ -68,6 +69,15 @@ const columnDataTypes = {
   TEXT: 'text',
 };
 
+const operators = [
+  { name: 'equals', value: '$eq', graphqlOp: '_eq' },
+  { name: 'not equals', value: '$ne', graphqlOp: '_neq' },
+  { name: '>', value: '$gt', graphqlOp: '_gt' },
+  { name: '<', value: '$lt', graphqlOp: '_lt' },
+  { name: '>=', value: '$gte', graphqlOp: '_gte' },
+  { name: '<=', value: '$lte', graphqlOp: '_lte' },
+];
+
 // eslint-disable-next-line no-useless-escape
 const createSQLRegex = /create\s*(?:|or\s*replace)\s*(view|table|function)\s*(?:\s*if*\s*not\s*exists\s*)?((\"?\w+\"?)\.(\"?\w+\"?)|(\"?\w+\"?))/g;
 
@@ -77,9 +87,26 @@ export const displayTableName = (table: Table) => {
   return isTable(table) ? <span>{tableName}</span> : <i>{tableName}</i>;
 };
 
+const processTableRowData = (
+  data: any,
+  config?: { originalTable: string; currentSchema: string }
+) => {
+  const { originalTable, currentSchema } = config!;
+  const rows =
+    data.data[
+      currentSchema === 'dbo'
+        ? originalTable
+        : `${currentSchema}_${originalTable}`
+    ];
+  return { estimatedCount: rows.length, rows };
+};
+
 export const mssql: DataSourcesAPI = {
   isTable,
   displayTableName,
+  operators,
+  getTableRowRequest,
+  processTableRowData,
   getFunctionSchema: () => {
     return '';
   },