+++ console/src/dataSources/services/postgresql/index.tsx
@@ -3,6 +3,7 @@ import { Table, TableColumn, ComputedField } from '../../types';
 import { QUERY_TYPES, Operations } from '../../common';
 import { PGFunction } from './types';
 import { DataSourcesAPI, ColumnsInfoResult } from '../..';
+import { getTableRowRequest } from './utils';
 import {
   getFetchTablesListQuery,
   fetchColumnTypesQuery,
@@ -354,6 +355,59 @@ const commonDataTypes = [
   },
 ];
 
+const operators = [
+  { name: 'equals', value: '$eq', graphqlOp: '_eq' },
+  { name: 'not equals', value: '$ne', graphqlOp: '_neq' },
+  { name: 'in', value: '$in', graphqlOp: '_in', defaultValue: '[]' },
+  { name: 'not in', value: '$nin', graphqlOp: '_nin', defaultValue: '[]' },
+  { name: '>', value: '$gt', graphqlOp: '_gt' },
+  { name: '<', value: '$lt', graphqlOp: '_lt' },
+  { name: '>=', value: '$gte', graphqlOp: '_gte' },
+  { name: '<=', value: '$lte', graphqlOp: '_lte' },
+  { name: 'like', value: '$like', graphqlOp: '_like', defaultValue: '%%' },
+  {
+    name: 'not like',
+    value: '$nlike',
+    graphqlOp: '_nlike',
+    defaultValue: '%%',
+  },
+  {
+    name: 'like (case-insensitive)',
+    value: '$ilike',
+    graphqlOp: '_ilike',
+    defaultValue: '%%',
+  },
+  {
+    name: 'not like (case-insensitive)',
+    value: '$nilike',
+    graphqlOp: '_nilike',
+    defaultValue: '%%',
+  },
+  { name: 'similar', value: '$similar', graphqlOp: '_similar' },
+  { name: 'not similar', value: '$nsimilar', graphqlOp: '_nsimilar' },
+
+  {
+    name: '~',
+    value: '$regex',
+    graphqlOp: '_regex',
+  },
+  {
+    name: '~*',
+    value: '$iregex',
+    graphqlOp: '_iregex',
+  },
+  {
+    name: '!~',
+    value: '$nregex',
+    graphqlOp: '_nregex',
+  },
+  {
+    name: '!~*',
+    value: '$niregex',
+    graphqlOp: '_niregex',
+  },
+];
+
 export const isColTypeString = (colType: string) =>
   ['text', 'varchar', 'char', 'bpchar', 'name'].includes(colType);
 
@@ -406,6 +460,15 @@ const getReferenceOption = (opt: string) => {
   }
 };
 
+const processTableRowData = (data: any) => {
+  let estimatedCount =
+    data.length > 1 && data[0].result > 1 && data.result[1].length
+      ? data[1].result[1][0]
+      : null;
+  estimatedCount =
+    estimatedCount !== null ? parseInt(data[1]?.result[1][0], 10) : null;
+  return { rows: data[0], estimatedCount };
+};
 const permissionColumnDataTypes = {
   boolean: ['boolean'],
   character: ['character', 'character varying', 'text', 'citext'],
@@ -501,6 +564,9 @@ export const postgres: DataSourcesAPI = {
   getEventInvocationInfoByIDSql,
   getDatabaseInfo,
   getTableInfo,
+  operators,
+  getTableRowRequest,
+  processTableRowData,
   getDatabaseVersionSql,
   permissionColumnDataTypes,
   viewsSupported: true,