+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 (Add entries here in the order of: server, console, cli, docs, others)
 
 - server: fix issue with parsing of remote schema list of input objects (fix #6584)
+- console: add browse rows for mssql tables (#805)
 
 
 ## v2.0.0-alpha.4